+++ packages/auth/src/features/authenticate/service.ts
@@ -42,7 +42,6 @@ import { UserAuditLog } from '@opencrvs/commons/events'
 import * as F from 'fp-ts'
 import {
   EncodedScope,
-  encodeScope,
   TokenUserType,
   TokenWithBearer
 } from '@opencrvs/commons/authentication'
@@ -248,51 +247,6 @@ export async function createRefreshToken(
 ): Promise<string> {
   const { familyId, jti } = await createFamily(userId)
   return signRefreshToken(userId, userType, familyId, jti)
-}
-
-type ActionConfirmationInput = {
-  eventId: UUID
-  actionId: UUID
-}
-
-type LegacyRecordValidationInput = {
-  recordId: UUID
-}
-
-export async function createTokenForActionConfirmation(
-  input: ActionConfirmationInput | LegacyRecordValidationInput,
-  userId: UUID,
-  userType: TokenUserType,
-  extraScopes: EncodedScope[] = []
-) {
-  return sign(
-    {
-      scope: [
-        encodeScope({ type: 'record.confirm-registration' }),
-        encodeScope({ type: 'record.reject-registration' }),
-        ...extraScopes
-      ],
-      eventId: 'eventId' in input ? input.eventId : undefined,
-      actionId: 'actionId' in input ? input.actionId : undefined,
-      recordId: 'recordId' in input ? input.recordId : undefined,
-      userType
-    },
-    cert,
-    {
-      subject: userId,
-      algorithm: 'RS256',
-      expiresIn: env.CONFIG_ACTION_CONFIRMATION_TOKEN_EXPIRY_SECONDS,
-      audience: [
-        'opencrvs:gateway-user',
-        'opencrvs:events-user',
-        'opencrvs:user-mgnt-user',
-        'opencrvs:auth-user',
-        'opencrvs:countryconfig-user',
-        'opencrvs:documents-user'
-      ],
-      issuer: JWT_ISSUER
-    }
-  )
 }
 
 export async function storeUserInformation(